Fixed LoadScript being not a BlockingQueuedConnection
This needs to be the case because if a script calls eval and in that eval a goto is
called, the user script needs to wait until the goto returns fully (or rather
aborts the wait and calls the new script). If we don't do this, the script may continue
when it is expected that we have already switched scenes. This is especially
important for EOS scripts as those are way more script heavy.